pyproject.toml
[project]
name = "auto-job-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.12.0",
    "colorlog>=6.10.1",
    "langchain>=1.2.10",
    "langchain-cohere>=0.5.0",
    "langchain-core>=1.2.16",
    "mem0ai>=1.0.4",
    "pypdf>=6.6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
